requirements.txt
aiofiles==25.1.0
aiogram==3.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.2
asyncpg==0.31.0
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
frozenlist==1.8.0
google-auth==2.48.0
google-auth-oauthlib==1.3.0
gspread==6.2.1
idna==3.11
magic-filter==1.0.12
multidict==6.7.1
numpy==2.4.2
oauthlib==3.3.1
pandas==3.0.1
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
yarl==1.22.0
